requirements.txt
Django==1.11.6
django-rest-framework==0.1.0
djangorestframework==3.7.0
factory-boy==2.9.2
Faker==0.8.5
ipaddress==1.0.18
python-dateutil==2.6.1
pytz==2017.2
six==1.11.0
Unidecode==0.4.21
